made all stat values doubles, started statGraph

+++ Process/quote.c
@@ -11,12 +11,12 @@ int getQuote(MYSQL_RES* res, QUOTES* quotes, int pos) {
 	MYSQL_ROW row;
 	row = mysql_fetch_row(res);
 	if (!row) return 0;
-	quotes->utime[pos] = atol(row[1]);
+	quotes->utime[pos] = atof(row[1]);
 	quotes->high[pos] = atof(row[2]);
 	quotes->low[pos] = atof(row[3]);
 	quotes->open[pos] = atof(row[4]);
 	quotes->close[pos] = atof(row[5]);
-	quotes->volume[pos] = atoi(row[6]);
+	quotes->volume[pos] = atof(row[6]);
 	return 1;
 }
 
@@ -27,12 +27,12 @@ QUOTES* getQuotes(MYSQL_RES* res) {
 	QUOTES* quotes = malloc(sizeof(QUOTES));
 	
 	quotes->count = count;
-	quotes->utime = malloc(sizeof(unsigned int) * count);
+	quotes->utime = malloc(sizeof(double) * count);
 	quotes->high = malloc(sizeof(double) * count);
 	quotes->low = malloc(sizeof(double) * count);
 	quotes->open = malloc(sizeof(double) * count);
 	quotes->close = malloc(sizeof(double) * count);
-	quotes->volume = malloc(sizeof(unsigned int) * count);
+	quotes->volume = malloc(sizeof(double) * count);
 
 	int pos = 0;
 	while(getQuote(res, quotes, pos++));
@@ -46,7 +46,7 @@ void printQuote(QUOTES* quotes, int place) {
 		return;
 	}
 
-	printf("%i | %f | %f | %f | %f | %i\n",
+	printf("%f | %f | %f | %f | %f | %f\n",
 		quotes->utime[place],
 		quotes->high[place],
 		quotes->low[place],
